Refactor some of the previous days code

+++ day01/main.cpp
@@ -88,7 +88,7 @@ namespace day01 {
     }
 
     /// Part One Solution
-    int partOne() {
+    int64_t partOne() {
         using namespace std::views;
 
         PuzzleInput input(inputFilename);
@@ -105,7 +105,7 @@ namespace day01 {
 
 
     /// Part Two Solution
-    int partTwo() {
+    int64_t partTwo() {
         using namespace std::views;
 
         PuzzleInput input(inputFilename);
@@ -125,6 +125,6 @@ namespace day01 {
 int main() {
     using namespace day01;
     test();
-    advent::run<int>(day, "One", partOne);
-    advent::run<int>(day, "Two", partTwo);
+    advent::run<int64_t>(day, "One", partOne);
+    advent::run<int64_t>(day, "Two", partTwo);
 }
