+++ day03/main.cpp
@@ -38,7 +38,7 @@ namespace day03 {
     }
 
     /// Part One Solution
-    int partOne() {
+    int64_t partOne() {
         using namespace std::views;
 
         PuzzleInput input(inputFilename);
@@ -48,7 +48,7 @@ namespace day03 {
 
 
     /// Part Two Solution
-    int partTwo() {
+    int64_t partTwo() {
         using namespace std::views;
 
         PuzzleInput input(inputFilename);
@@ -62,6 +62,6 @@ namespace day03 {
 int main() {
     using namespace day03;
     test();
-    advent::run<int>(day, "One", partOne);
-    advent::run<int>(day, "Two", partTwo);
+    advent::run<int64_t>(day, "One", partOne);
+    advent::run<int64_t>(day, "Two", partTwo);
 }
